a step-through debugging feature
Right-mouse clicking a Floor Tile starts to run the pathfinding Algorithm from the Monster’s position, but does not move any characters. Instead, it starts stepping through the algorithm slowly over the course of about a minute, allowing one to see intermediary steps as the algorithm runs. Users should be able to hold or tap spacebar to speed up the step-through. Then, pressing Enter should cause the character to actually follow the path, completing the motion.
○ All tile costs currently found by the pathfinding algorithm are displayed on each tile